+++ src/components/post/Post.jsx
@@ -6,7 +6,7 @@ import { collection, addDoc, doc, deleteDoc, setDoc, query, where, onSnapshot, s
 import "../../css/components/Post.css";
 import AppContext from "../../context/AppContext";
 import InputImage from "../image/InputImage";
-import { UploadImageToStorage } from "../image/UploadImageToStorage";
+import { SetThumbnail } from "../image/SetThumbnail";
 
 //記事の編集が押されたら（EditPost.jsx）から情報を{EditPost}に渡してる
 function Post({ EditPost }) {
@@ -19,8 +19,6 @@ function Post({ EditPost }) {
   const [fileObject, setFileObject] = useState(null);
   const [newProfilePicture, setNewProfilePicture] = useState(''); //変更後のプロフィール画像
   let docRef = "";
-  const [thumbnailStorageUrl, setThumbnailStorageUrl] = useState('');
-  const defaultThumbnailUrl = "https://firebasestorage.googleapis.com/v0/b/blogapp-c1052.appspot.com/o/postImages%2Fdefault%2FnoImage.PNG?alt=media&token=c7febf2c-b45e-4469-b1ce-2b232f8bd56d"
 
   // Firebaseの中にあるpostsのフィールドから、ユーザーの投稿記事を取得
   useEffect(() => {
@@ -58,7 +56,7 @@ function Post({ EditPost }) {
       alert("下書きに保存できるのは5個までです。");
     }
 
-    setThumbnail(); //サムネイル画像の設定
+    SetThumbnail(fileObject, docRef); //サムネイル画像の設定
   }
 
   // 編集画面で下書き移動ボタンを押したら発火
@@ -121,45 +119,13 @@ function Post({ EditPost }) {
       }
     }
 
-    setThumbnail(); //サムネイル画像の設定
+    SetThumbnail(fileObject, docRef); //サムネイル画像の設定
     setTitle("");
     setContent("");
     navigate(`/${user.uid}`); // "/mypage"に移動
   }
 
-  //サムネイル画像の設定
-  const setThumbnail = async () => {
-    let thumbnailStorageUrl = "";
-    //サムネイル画像が選択された場合
-    if (fileObject) {
-      //ストレージ保存先のパスを指定
-      const storageFilePath = "postImages/" + docRef.id + "/thumbnail/";
-      //ストレージに画像を保存、URLを取得する（非同期処理でアップロード完了を待つ）
-      thumbnailStorageUrl = await UploadImageToStorage(fileObject, storageFilePath);
-    }
-    //選択されていない場合
-    else {
-      // ドキュメントを取得してthumbnailUrlが存在するか確認
-      const docSnapshot = await getDoc(docRef);
-      if (docSnapshot.exists()) {
-        const data = docSnapshot.data();
-        // thumbnailUrlが既に存在する場合は処理を抜ける
-        //下書きから投稿する際にdefaultThumbnailUrlに変更されてしまうため
-        if (data.thumbnailUrl) {
-          return
-        }
-      }
 
-      //NO IMAGE画像を設定
-      thumbnailStorageUrl = defaultThumbnailUrl;
-    }
-    // ドキュメントの参照を取得し、thumbnailUrlを追加
-    const postDocRef = doc(db, "posts", docRef.id);
-    //ポストのstorageにサムネイル画像のurlを追加する
-    await updateDoc(postDocRef, {
-      thumbnailUrl: thumbnailStorageUrl,
-    });
-  }
 
   return (
     <>
